Functions should either be inline, or extern not both
When using GCC 6, the default is to compile for the "gnu1x" standard:
https://gcc.gnu.org/onlinedocs/gcc/C-Dialect-Options.html
which means that the standard used is a variant of C11, so a version
newer than C99.

From http://stackoverflow.com/questions/216510/extern-inline, for C99:
"
extern inline: like GNU "inline": externally visible code is emitted, so
at most one translation unit can use this.
"

This means that the code will be externally visible for each object
compiled that includes it. This makes it impossible to link the objects
generated as the function will be defined multiple times.

Making the function "inline" rather than "extern inline" means that it
will be copied in each of the modules, at a small cost in size, but
makes it possible to link the resulting binary.