Disable motion compression
By default, motion events are compressed to avoid waking up applications
for tiny changes, especially for sub-pixel mouse movements. However, we
do need all the events if we want the "Crayon optique" to follow the
mouse.

This functionality is only available with GTK+ 3.12, so won't work as
expected under Ubuntu 14.04's GTK+ 3.10.8.