Disable warnings as errors
And remove deprecation warning flags. Errors and deprecation errors
should only be enabled by the developer for their own sake. When
distributing code, if that particular code used to work, we might break
it with a newer library, or new GCC warnings, but the application would
still work just as well.