linux: Remove Version entry from desktop file
The version field is for the version of the desktop entry specification
that the desktop file conforms to, not the version of the software:
https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html#recognized-keys

"
Version of the Desktop Entry Specification that the desktop entry conforms
with. Entries that confirm with this version of the specification should
use 1.1. Note that the version field is not required to be present.
"