Fix floppy disk loading on the command-line (again)
sap_IsSap(), hfe_IsHfe() and fd_IsFd () don't return a simple TRUE/FALSE,
but negative values on error, > 0 if protected, and 0 if not protected.
The committer of the previous fix (me) tested with a write-protected
floppy, and checking for "== TRUE" fixed his problem, but we should really
be checking for the absence of errors instead.

Fixes: "Correct floppy disk loading on the command-line"