Fix sound cracks
The problem was in two parts. Frist part was introduced when converting
original Allegro code to SDL. The SDL version did not pushed a sound
frame when last_index didn't move (i.e, no PutSoundByte call was received).
Second part(that let to the first part as a 'fix') was doing a time-based
sync sound being on or off. The right thing to do was to sync on sound
when sound was on (wait for the buffer to have no more than one frame
stacked) and on time(TEO_MICROSECONDS_PER_FRAME) when sound is off.